requirements.txt
pytest==2.9.0
pytest-allure-adaptor==1.7.10
requests==2.18.4
ws4py>=0.3.5